+++ packages/core/src/Popover.js
@@ -24,12 +24,15 @@ export const BEM = {
   container: ROOT_BEM.element('container'),
 };
 
+const POPOVER_PADDING = 24;
+
 function Popover({
   onClick,
   // from anchored()
   placement,
   arrowStyle,
   nodeRef,
+  remainingSpace,
   // from closable()
   onInsideClick,
   // React props
@@ -39,6 +42,13 @@ function Popover({
 }) {
   const bemClass = BEM.root.modifier(placement);
   const rootClassName = classNames(bemClass.toString(), className);
+  /**
+   * The `remainingSpace` is the space for whole popover.
+   * What we want here is to always show keep `remainingSpace === popoverHeight`
+   * The `maxHeight` is for `BEM.container`, which doesn't include root class padding.
+   * So we need to minus POPOVER_PADDING here.
+   */
+  const maxHeight = remainingSpace ? remainingSpace - POPOVER_PADDING : undefined;
 
   const handleWrapperClick = (event) => {
     onInsideClick(event);
@@ -55,7 +65,10 @@ function Popover({
         {...otherProps}
       >
         <span className={BEM.arrow} style={arrowStyle} />
-        <div className={BEM.container}>
+        <div
+          className={BEM.container}
+          style={{ maxHeight }}
+        >
           {children}
         </div>
       </div>
@@ -68,6 +81,7 @@ Popover.propTypes = {
   placement: anchoredPropTypes.placement,
   arrowStyle: anchoredPropTypes.arrowStyle,
   nodeRef: anchoredPropTypes.nodeRef,
+  remainingSpace: anchoredPropTypes.remainingSpace,
   onInsideClick: PropTypes.func.isRequired,
 };
 
@@ -76,6 +90,7 @@ Popover.defaultProps = {
   placement: ANCHORED_PLACEMENT.BOTTOM,
   arrowStyle: {},
   nodeRef: undefined,
+  remainingSpace: undefined,
 };
 
 export { Popover as PurePopover };